Refactor auth service into modules and fix banner section
