+++ views/homepage.ejs
@@ -13,7 +13,6 @@
 </head>
 
 <body>
-
   <header>
     <h1 class="logo">Rate Reads</h1>
     <div class="toggle-menu">
@@ -32,21 +31,22 @@
 
   <section class="banner">
     <form class="search-book-form">
-      <label for="book-name-input" class="search-label">
-        <h2>Discover your new favorite book</h2>
-      </label>
-      <div class="search-bar">
-        <input type="text" id="book-name-input" placeholder="Explore titles">
-        <button class="search-button"><i class="fa fa-search"></i></button>
+      <div class="form-wrapper">
+        <label for="book-name-input" class="search-label">
+          <h2>Discover your new favorite book</h2>
+        </label>
+        <div class="search-bar">
+          <input type="text" id="book-name-input" placeholder="Explore titles">
+          <button class="search-button"><i class="fa fa-search"></i></button>
+        </div>
       </div>
     </form>
   </section>
 
-
-
   <h2>Login Form</h2>
   <form class="login-form">
     <label for="username-input">Enter username</label>
+    <br>
     <input type="text" id="username-input">
 
     <label for="password-input">Enter password</label>
@@ -55,8 +55,9 @@
     <button>Submit</button>
   </form>
 
+
   <script src="/app.js"></script>
-  <script src="/navbar.js"></script>
+  <script type="module" src="/navbar.js"></script>
 </body>
 
 </html>
